Osculp32k, Xosc32k: Remove support for a write lock
Write lock is a difficult mechanism to model. In case of a non-POR reset with
a write lock engaged, clocking initialization procedure becomes basically
undefined and types do not reflect the correct state of a hardware. Counter
increment on an `Enabled` type is not enough to model this behaviour correctly.