clock/v2: Move `Enabled` from types module to top-level
Now that `Counter` and its associated traits have been moved to the
`typelevel` module, the `types` module only contains the `Enabled`
wrapper type. Migrate this to the top-level of the `clock::v2` module,
where it fits better. Leave the empty `types` module for now, since it
contains documentation we may want to salvage in the future.