Xosc: Simplify the `xosc` module
Modify the `Mode` trait so that the type parameter can be removed from
`CrystalMode`. This eliminates redundancy when specifying the full type
of an `Xosc`. Also add a `configure` function to the `Mode` trait, so
that the `Xosc::enable` function can be generic. Rename the variants of
`CrystalCurrent` to remove the frequency range. Add this information to
the documentation, and refer to the datasheet for optimizing power
usage. Adjust the way a frequency, in `Hertz`, is converted to a
`CrystalCurrent`.