clock/v2: Update 1 kHz, 32 kHz and RTC clocks
Refactor the slow osccilators to mimic the structure in the rest of the
`clock` module. Rather than use extra type parameters to represent the
enabled or disabled state of the 1 kHz and 32 kHz outputs, instead
create a `Base` clock that can only be used to enable downstream 1 kHz
and 32 kHz clocks. With this approach, the information is encoded in the
type parameter `N` of `Enabled<Clock, N>`, rather than dedicated type
parameters `X`` and `Y`. This approach improves consistency with the
rest of the `clock` module.