Xosc: Update `xosc::Mode` and `xosc::CrystalCurrent`
Use the `Mode` trait to map to the `XOut` type, allowing it to be
removed from the `CrystalConfig`. Merge `CrystalConfig` and
`CrystalMode`. Now `CrystalMode` is no longer generic. Remove any
abstraction of behavior from the `Mode` trait. Instead, have it
represent an interface to the underlying configuration data. For
`ClockMode`, the underlying values are the reset values for the
corresponding fields.

Remove the implementation of `From<Hertz>` for `CrystalMode`, which also
removes any possibility of failure. Instead, for the user to choose a
`CrystalCurrent` manually, when calling `from_crystal`. Add a `Zero`
variant to `CrystalCurrent` to represent the default values of `IMULT`
and `IPTAT`.