clock/2: Overhaul preset clock configurations
Reimplement the preset clock configurations as real functions and
structs, rather than macros. The result has clear benefits, because the
preset configurations are now checked at compile-time for the HAL,
unlike the macro versions, which are not checked until compile-time for
users.

To implement these changes, modify the way `PclkId` types are generated
and used to implemen the `pclk::Tokens` struct. The new approach allows
better reuse of the `PclkId` types throughout the HAL, including the new
`clock::configurations` module.