clock/v2: Remove documentation warnings
Remove warnings for missing documentation. This is only temporary, while
the code is still in flux. It will eventually be added back. Also remove
a `TODO` that was surpressing this warning for the `rtcosc` module. It
is better to leave off the `TODO`s and use the compiler to find missing
documentation later.