gclk: Refactor handling of GCLK_IO and update documentation
`GclkIn` was a 1:1 clock, but it used `Enabled` and `Source` as if it
where a 1:N clock. Instead, remove `GclkIn` completely and modify `Gclk`
to use `from_source` or `from_pin` constructors.

Allow multiple `GclkOut`s per `Gclk`, since this is allowed by the
hardware. Doing so removes the need for any `GclkOutToken`s.

Consolidate the `gclkio` module with the `gclk` module, because it has
been reduced to only a single trait and type.

Update the documentation and code throughout to match the changes.